Adding the ability to delete all attestation keys
This feature primarily supports better testing, by making sure the
system is completely cleaned up after running tests. There are potential
other use cases for this, such as deleting attestation keys that haven't
yet been included in a certificate signing request, if there are
compatibility issues in CBOR formats during an upgrade of the KM HAL. It
would be simpler and more robust to simply delete and regenerate keys on
the upgraded instance than attempt to upgrade the CBOR formats.

Test: keystore2_test

Change-Id: I8a636a47f54dc9265f5f503ac59546aa717ea61b