Modifying calls to RemotelyProvisionedComponent HAL
This change just updates the calls into the underlying HAL interface for
remote provisioning to match the AIDL change introduced in there. The
change in question is just part of the portion of the remote
provisioning flow wherein keystore2 acts as a relay between the server
and the HAL, so there is essentially no functionality change here.

Test: atest keystore2_test
Change-Id: Ibe376f7d156d93cdff73a459222c0d7085b4d7df