Keystore 2.0: Reinstate the default busy handler.
The default busy handler leads to a semi deadlock when used in
conjunction with an in-memory database. That is, the busy handler would
time out because a mutex was held by a thread trying to acquire the file
lock.

The in-memory database was removed from keystore2, so the default busy
handler may be reinstated.

Test: keystore2_test
Bug: 184006658
Change-Id: Idf3a50250342b9eb677b460074dfc6ee7df73964