Keystore 2.0: Extend the functionality of the Vpn profile store.
It turns out there are more clients that use Keystore in a creative
way. This patch renames the VpnProfileStore to LegacyKeystore and
extends the functionality such that it allows access to all blobs with
alias prefixes that were not known to Keystore. It also brings back the
option to specify a uid argument. Specifically, for AID_SYSTEM to
manipulate the WIFI namespace.

Ignore-AOSP-First: No mergepath from AOSP

Test: TBD
Bug: 191373871
Change-Id: Iaf81e7ccaee3c09a465dcec0fd5899b781c31db5