Optimize dependencies in rkp_factory_extraction_tool
The binary depended on a number of shared libs, but now uses static
linking where possible.

Stop linking to libvintf, which may not be available in a factory
image (it's not vendor-available).

Mark rkp_factory_extraction_tool as a vendor binary, allowing it to
be included in factory images.

Bug: 191301285
Test: Manually run tool on device
Ignore-AOSP-First: This is a port from AOSP, as there's no merge path.
Change-Id: I1747f809ffd2d3d605c9f4ee81e9f40cc3a189c4
Merged-In: I1747f809ffd2d3d605c9f4ee81e9f40cc3a189c4